+++ code/nn.py
@@ -103,8 +103,8 @@ class NeuralNetwork:
 			morph_b_prev, morph_b, morph_b_next], mode='concat')
 		morph = Dense(64, init='uniform', activation='relu')(morph)
 		
-		lemma_a = Input(shape=(1,), dtype='uint32')
-		lemma_b = Input(shape=(1,), dtype='uint32')
+		lemma_a = Input(shape=(1,), dtype='uint16')
+		lemma_b = Input(shape=(1,), dtype='uint16')
 		lemma_embed = Embedding(vocab_sizes['lemmas'], 256, input_length=1)
 		lemmas = merge([
 			Flatten()(lemma_embed(lemma_a)),
